refactor(saga): Phase 4 relocate SagaTelemetry + InternalOutboxEndpoints
- SagaTelemetry moves to Infrastructure/Observability/ under namespace
  Saga.Service.Infrastructure.Observability; top-level Observability/
  folder deleted. Activity-source + meter + counter/histogram names
  preserved verbatim.
- InternalOutboxEndpoints moves to Infrastructure/Outbox/ under namespace
  Saga.Service.Infrastructure.Outbox; RequireService policy gate
  preserved. Endpoints/ still holds OperatorSagaEndpoints.cs (Phase 8
  dissolves it).
