+++ saga-microservice/Saga.Service/IntegrationEvents/EventHandlers/OrderSagaReplyProcessor.cs
@@ -3,8 +3,8 @@ using Saga.Service.Contracts.Integration.InboundEvents;
 using Saga.Service.Domain;
 using Saga.Service.Domain.Abstractions;
 using Saga.Service.Domain.OrderSaga;
+using Saga.Service.Infrastructure.Observability;
 using Saga.Service.Infrastructure.Reaper;
-using Saga.Service.Observability;
 
 namespace Saga.Service.IntegrationEvents.EventHandlers;
