+++ saga-microservice/Saga.Service/IntegrationEvents/EventHandlers/RefundRequestedEventHandler.cs
@@ -5,7 +5,7 @@ using Saga.Service.Contracts.Integration.InboundEvents;
 using Saga.Service.Domain;
 using Saga.Service.Domain.RefundSaga;
 using Saga.Service.Infrastructure.Data.EntityFramework;
-using Saga.Service.Observability;
+using Saga.Service.Infrastructure.Observability;
 
 namespace Saga.Service.IntegrationEvents.EventHandlers;
